# fix extra heap allocations when detector is disabled