## Summary
I found deadlock detector has one allocation overhead per locker call b/c it passes `Lock()` method to the `lock` helper, and it escapes.

## Benchmark

Before:
```
go test ./ -run ^$ -bench BenchmarkRWMutexAlloc -benchmem -memprofile=mem.pprof
goos: darwin
goarch: arm64
pkg: github.com/algorand/go-deadlock
BenchmarkRWMutexAlloc-10    	32553236	        36.69 ns/op	      16 B/op	       1 allocs/op
PASS
ok  	github.com/algorand/go-deadlock	2.273s
```
After:
```
go test ./ -run ^$ -bench BenchmarkRWMutexAlloc -benchmem -memprofile=mem.pprof
goos: darwin
goarch: arm64
pkg: github.com/algorand/go-deadlock
BenchmarkRWMutexAlloc-10    	45803848	        26.02 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/algorand/go-deadlock	2.201s
```